Combine Lumen & Laravel service providers
